Workaround winit bug where events execute immediately (#4)
Winit on web has a bug where the EventLoopProxy will immediately call handlers for a custom event in send_event instead of scheduling it for the next event tick. This causes double-borrows of our MutableAppContext, which hard crashes the app. This is a bandaid fix which pushes those calls on the microtask queue instead.